requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1
ryugraph>=25.9.1
openai>=1.0.0
litellm>=1.0.0
numpy>=1.24.0
networkx>=3.0
scipy>=1.10.0
tenacity>=8.0.0
rank-bm25>=0.2.2
requests>=2.31.0
google-genai>=0.2.0
pandas>=2.0.0
google-adk>=1.18.0